Make compatible_language_from follow the RFC 2616 selection algorithm
The change in commit 92b276c92db8dae14c51169d6567c2f7a2563141 is wrong
because it reverses the previous change to obey the RFC 2616 algorithm.

Specifically RFC 2616 says that a request for en-GB will only match
with en-GB or something that starts with en-GB and it will not match
against en with no suffix.

The test failures were because the test was wrong, not because the
code was wrong. This commit restores the previous code and fixes the
tests.