Default `Accept-Language` to `I18n.default_locale`
Currently AutoLocale tries to set `I18n.locale` to nil when the
Accept-Language header is missing, which leads to "nil is not a valid
locale" error when `I18n.config.enforce_available_locales` is
true (default).

This commit fixes the problem by defaulting the header value to
`I18n.default_locale`.